Massive refactor to allow cycle-level timing (#7)
In order to accommodate timing with other components such as the PPU, the CPU's step function now accepts a tick handler that is executed during every individual read/write.  The exact timing for each instruction is very likely not correct as of this commit, timing tests and tweaks are to follow.